Add files via upload
This project looked to create a CNN model to detect Pneumonia in X-rays. The case goal was to fundraise imaginary money to implement the model in a third-world country. We performed a cost-benefit analysis of the CNN to decipher whether such a project should be pursued. I am not able to upload the code, but this presentation represents our process and findings.